+++ packages/about-you/composables/nuxt/README.MD
@@ -1,0 +1,13 @@
+# Example config
+```js
+export const config = {
+  api: {
+    host: 'https://boston.backbone-api.demo.aboutyou.cloud/v1/',
+    auth: {
+      username: 'aboutyou',
+      password: 'OmNErAb96Y5Qn75SFhXr'
+    }
+  },
+  imgUrl: 'https://mt1.dam.demo.aboutyou.cloud/boston',
+};
+```
